I decided that there is no real value to adding groovy
We should really use it when needed. Right now the size of the artifacts
is bigger, there is more quirks around joint compilation, and it's just
generally not worth it to add a new language to an old, fully developed
system. I'd rather just maintain, upgrade, and do whatever is necessary
to keep the code clean and performant.