fix bug in which cannot exit battle due to missing fields
They were removed because it was thought they're not referenced,
but they were in fact referenced from beanshell scripts. Extra care must
be taken because removing anything because the environment is not only
compiled java but also beanshell scripts that are only applicable during
runtime and do not report errors or exceptions.