incorrect workaround removed
The problem is not multi-threaded or java at all. The problem is the
specific hardware configuration on my machine which grabs exclusive
hardware ALSA access instead of software mixer, thus being unable to
play sounds concurrently on Linux.

Maybe one workaround to attempt is to share the SourceDataLine. Still
WIP and experimenting